+++ Framework/Interface/IShaderManager.hpp
@@ -29,7 +29,7 @@ namespace My {
         virtual bool InitializeShaders() = 0;
         virtual void ClearShaders() = 0;
 
-        virtual intptr_t GetDefaultShaderProgram(DefaultShaderIndex index) = 0;
+        virtual int32_t GetDefaultShaderProgram(DefaultShaderIndex index) = 0;
     };
 
     extern IShaderManager* g_pShaderManager;
